Made extensive changes to ito_process file
Now instead of multiple different solving functions for each method,
there will only be a single solving function with an option included
to switch between different methods. Each of these methods are
encapsulated in a different function that executes a single iteration
of each method. Different distributions are also encapsulated in their
own functions.

An error exists in the file that will need to be resolved in the
strong_taylor_iteration function.